Separate custom feature shape validation
