+++ src/lib/customFeatureValidation.ts
@@ -1,6 +1,6 @@
 import type { CustomPolygonFeature } from '../types/customProducts';
 import { CUSTOM_PRODUCT_LIMITS } from '../types/customProducts';
-import { getCustomGeometryDimension, isCustomPolygonGeometry } from './customGeometry';
+import { hasValidCustomFeatureShape } from './customFeatureShape';
 
 const isRecord = (value: unknown): value is Record<string, unknown> =>
   Boolean(value && typeof value === 'object' && !Array.isArray(value));
@@ -11,29 +11,6 @@ const hasOnlyKeys = (value: Record<string, unknown>, keys: readonly string[]): b
 const isBoundedText = (value: unknown): value is string => {
   if (typeof value !== 'string' || value.trim() !== value) return false;
   return value.length > 0 && value.length <= CUSTOM_PRODUCT_LIMITS.labelLength;
-};
-
-const isBoundingBox = (value: unknown, dimension: number): boolean => {
-  if (!Array.isArray(value) || value.length !== dimension * 2) return false;
-  if (!value.every((coordinate) => typeof coordinate === 'number' && Number.isFinite(coordinate))) return false;
-  return value.slice(0, dimension).every((minimum, index) => minimum <= value[index + dimension]);
-};
-
-const isOptionalFeatureId = (value: unknown): boolean => {
-  if (value === undefined || typeof value === 'string') return true;
-  return typeof value === 'number' && Number.isFinite(value);
-};
-
-const isFeatureType = (value: Record<string, unknown>): boolean => value.type === 'Feature';
-
-const hasFeatureProperties = (value: Record<string, unknown>): boolean => isRecord(value.properties);
-
-const hasValidShape = (value: Record<string, unknown>): boolean => {
-  if (!hasOnlyKeys(value, ['type', 'id', 'geometry', 'properties', 'bbox'])) return false;
-  if (!isFeatureType(value) || !isCustomPolygonGeometry(value.geometry)) return false;
-  if (!hasFeatureProperties(value) || !isOptionalFeatureId(value.id)) return false;
-  if (value.bbox === undefined) return true;
-  return isBoundingBox(value.bbox, getCustomGeometryDimension(value.geometry));
 };
 
 const hasValidProperties = (properties: Record<string, unknown>): boolean => {
@@ -59,7 +36,7 @@ export const isCustomPolygonFeature = (
   layerId?: string,
   categoryIds?: ReadonlySet<string>,
 ): value is CustomPolygonFeature => {
-  if (!isRecord(value) || !hasValidShape(value)) return false;
+  if (!isRecord(value) || !hasValidCustomFeatureShape(value)) return false;
   const properties = value.properties as Record<string, unknown>;
   return hasValidProperties(properties) && matchesOwner(properties, layerId, categoryIds);
 };
